Movido tag para o mod_shared. Corrigido migrations
